+++ Nyong/week7/nikecompose/app/src/main/java/com/example/nikecompose/core/component/NikeBottomBar.kt
@@ -1,0 +1,58 @@
+package com.example.nikecompose.core.component
+
+import androidx.compose.foundation.layout.size
+import androidx.compose.material3.Icon
+import androidx.compose.material3.NavigationBar
+import androidx.compose.material3.NavigationBarItem
+import androidx.compose.material3.NavigationBarItemDefaults
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.nikecompose.navigation.BottomRoute
+
+@Composable
+fun NikeBottomBar(
+    items: List<BottomRoute>,
+    currentRoute: String?,
+    onItemClick: (BottomRoute) -> Unit
+) {
+    NavigationBar(
+        containerColor = Color.White,
+        tonalElevation = 0.dp
+    ) {
+        items.forEach { item ->
+            val selected = currentRoute == item.route
+            val iconColor = if (selected) Color.Black else Color.Gray
+
+            NavigationBarItem(
+                selected = selected,
+                onClick = { onItemClick(item) },
+                icon = {
+                    Icon(
+                        painter = painterResource(id = item.iconResId),
+                        contentDescription = item.label,
+                        modifier = Modifier.size(22.dp),
+                        tint = iconColor
+                    )
+                },
+                label = {
+                    Text(
+                        text = item.label,
+                        fontSize = 10.sp
+                    )
+                },
+                colors = NavigationBarItemDefaults.colors(
+                    selectedIconColor = Color.Black,
+                    selectedTextColor = Color.Black,
+                    unselectedIconColor = Color.Gray,
+                    unselectedTextColor = Color.Gray,
+                    indicatorColor = Color.Transparent
+                )
+            )
+        }
+    }
+}
