+++ Nyong/week7/nikecompose/app/src/main/java/com/example/nikecompose/feature/main/MainActivity.kt
@@ -4,61 +4,24 @@ import android.os.Bundle
 import androidx.activity.ComponentActivity
 import androidx.activity.compose.setContent
 import androidx.activity.enableEdgeToEdge
-import androidx.compose.foundation.Image
-import androidx.compose.foundation.background
-import androidx.compose.foundation.layout.Arrangement
-import androidx.compose.foundation.layout.Box
-import androidx.compose.foundation.layout.Column
-import androidx.compose.foundation.layout.PaddingValues
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.WindowInsets
-import androidx.compose.foundation.layout.aspectRatio
 import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
-import androidx.compose.foundation.layout.navigationBars
 import androidx.compose.foundation.layout.padding
-import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.layout.windowInsetsBottomHeight
-import androidx.compose.foundation.lazy.LazyColumn
-import androidx.compose.foundation.lazy.LazyRow
-import androidx.compose.foundation.lazy.grid.GridCells
-import androidx.compose.foundation.lazy.grid.LazyVerticalGrid
-import androidx.compose.foundation.lazy.grid.items
-import androidx.compose.foundation.lazy.items
-import androidx.compose.material3.Button
-import androidx.compose.material3.ButtonDefaults
-import androidx.compose.material3.Icon
-import androidx.compose.material3.NavigationBar
-import androidx.compose.material3.NavigationBarItem
-import androidx.compose.material3.NavigationBarItemDefaults
 import androidx.compose.material3.Scaffold
-import androidx.compose.material3.Tab
-import androidx.compose.material3.TabRow
-import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.mutableIntStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.setValue
-import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.text.style.TextAlign
-import androidx.compose.ui.unit.dp
-import androidx.compose.ui.unit.sp
 import androidx.navigation.NavGraph.Companion.findStartDestination
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import androidx.navigation.compose.currentBackStackEntryAsState
 import androidx.navigation.compose.rememberNavController
-import com.example.nikecompose.R
-import com.example.nikecompose.ui.theme.NikeComposeTheme
+import com.example.nikecompose.core.component.NikeBottomBar
+import com.example.nikecompose.feature.cart.presentation.CartScreen
+import com.example.nikecompose.feature.home.presentation.HomeScreen
+import com.example.nikecompose.feature.profile.presentation.ProfileScreen
+import com.example.nikecompose.feature.shop.presentation.ShopScreen
+import com.example.nikecompose.feature.wishlist.presentation.WishlistScreen
 import com.example.nikecompose.navigation.BottomRoute
+import com.example.nikecompose.ui.theme.NikeComposeTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
@@ -73,58 +36,6 @@ class MainActivity : ComponentActivity() {
     }
 }
 
-data class ProductItem(
-    val id: Int,
-    @DrawableRes val imageResId: Int,
-    val name: String,
-    val subTitle: String,
-    val colors: String,
-    val price: String,
-    val isLiked: Boolean = false,
-    val badge: String? = null
-)
-
-val productList = listOf(
-    ProductItem(
-        id = 1,
-        imageResId = R.drawable.img_shoe_1,
-        name = "Nike Everyday Plus Cushioned",
-        subTitle = "Training Ankle Socks (6 Pairs)",
-        colors = "5 Colours",
-        price = "US$10",
-        isLiked = true
-    ),
-    ProductItem(
-        id = 2,
-        imageResId = R.drawable.img_shoe_2,
-        name = "Nike Elite Crew",
-        subTitle = "Basketball Socks",
-        colors = "7 Colours",
-        price = "US$16",
-        isLiked = false
-    ),
-    ProductItem(
-        id = 3,
-        imageResId = R.drawable.img_shoe_1,
-        name = "Nike Air Force 1 '07",
-        subTitle = "Women's Shoes",
-        colors = "5 Colours",
-        price = "US$115",
-        isLiked = false,
-        badge = "BestSeller"
-    ),
-    ProductItem(
-        id = 4,
-        imageResId = R.drawable.img_shoe_2,
-        name = "Jordan ENike Air Force 1 '07ssentials",
-        subTitle = "Men's Shoes",
-        colors = "2 Colours",
-        price = "US$115",
-        isLiked = false,
-        badge = "BestSeller"
-    )
-)
-
 @Composable
 fun NikeComposeApp() {
     val navController = rememberNavController()
@@ -138,8 +49,7 @@ fun NikeComposeApp() {
     )
 
     Scaffold(
-        modifier = Modifier.Companion.fillMaxSize(),
-        containerColor = Color.Companion.White,
+        modifier = Modifier.fillMaxSize(),
         bottomBar = {
             NikeBottomBar(
                 items = bottomItems,
@@ -159,7 +69,7 @@ fun NikeComposeApp() {
         NavHost(
             navController = navController,
             startDestination = BottomRoute.Home.route,
-            modifier = Modifier.Companion
+            modifier = Modifier
                 .fillMaxSize()
                 .padding(innerPadding)
         ) {
@@ -193,452 +103,5 @@ fun NikeComposeApp() {
                 ProfileScreen()
             }
         }
-    }
-}
-
-@Composable
-fun NikeBottomBar(
-    items: List<BottomRoute>,
-    currentRoute: String?,
-    onItemClick: (BottomRoute) -> Unit
-) {
-    NavigationBar(
-        containerColor = Color.Companion.White,
-        tonalElevation = 0.dp
-    ) {
-        items.forEach { item ->
-            val selected = currentRoute == item.route
-            val iconColor = if (selected) Color.Companion.Black else Color.Companion.Gray
-
-            NavigationBarItem(
-                selected = selected,
-                onClick = { onItemClick(item) },
-                icon = {
-                    Icon(
-                        painter = painterResource(id = item.iconResId),
-                        contentDescription = item.label,
-                        modifier = Modifier.Companion.size(22.dp),
-                        tint = iconColor
-                    )
-                },
-                label = {
-                    Text(
-                        text = item.label,
-                        fontSize = 10.sp
-                    )
-                },
-                colors = NavigationBarItemDefaults.colors(
-                    selectedIconColor = Color.Companion.Black,
-                    selectedTextColor = Color.Companion.Black,
-                    unselectedIconColor = Color.Companion.Gray,
-                    unselectedTextColor = Color.Companion.Gray,
-                    indicatorColor = Color.Companion.Transparent
-                )
-            )
-        }
-    }
-}
-
-@Composable
-fun HomeScreen() {
-    LazyColumn(
-        modifier = Modifier.Companion
-            .fillMaxSize()
-            .background(Color.Companion.White),
-        contentPadding = PaddingValues(top = 32.dp, bottom = 24.dp)
-    ) {
-        item {
-            Column(
-                modifier = Modifier.Companion.padding(horizontal = 24.dp)
-            ) {
-                Text(
-                    text = "Discover",
-                    fontSize = 26.sp,
-                    fontWeight = FontWeight.Companion.Bold,
-                    color = Color.Companion.Black
-                )
-
-                Spacer(modifier = Modifier.Companion.height(4.dp))
-
-                Text(
-                    text = "9월 4일 목요일",
-                    fontSize = 13.sp,
-                    color = Color.Companion.Gray
-                )
-
-                Spacer(modifier = Modifier.Companion.height(28.dp))
-
-                Image(
-                    painter = painterResource(id = R.drawable.img_home_logo),
-                    contentDescription = "홈 배너 이미지",
-                    modifier = Modifier.Companion.fillMaxWidth(),
-                    contentScale = ContentScale.Companion.FillWidth
-                )
-
-                Spacer(modifier = Modifier.Companion.height(24.dp))
-
-                Text(
-                    text = "What's new",
-                    fontSize = 12.sp,
-                    fontWeight = FontWeight.Companion.Bold,
-                    color = Color.Companion.Gray
-                )
-
-                Spacer(modifier = Modifier.Companion.height(4.dp))
-
-                Text(
-                    text = "나이키 최신 상품",
-                    fontSize = 20.sp,
-                    fontWeight = FontWeight.Companion.Bold,
-                    color = Color.Companion.Black
-                )
-
-                Spacer(modifier = Modifier.Companion.height(16.dp))
-            }
-        }
-
-        item {
-            LazyRow(
-                contentPadding = PaddingValues(horizontal = 24.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp)
-            ) {
-                items(
-                    items = productList.take(2),
-                    key = { product -> product.id }
-                ) { product ->
-                    HomeProductCard(product = product)
-                }
-            }
-        }
-    }
-}
-
-@Composable
-fun HomeProductCard(
-    product: ProductItem
-) {
-    Column(
-        modifier = Modifier.Companion.width(230.dp)
-    ) {
-        Image(
-            painter = painterResource(id = product.imageResId),
-            contentDescription = product.name,
-            modifier = Modifier.Companion
-                .fillMaxWidth()
-                .height(230.dp)
-                .background(Color(0xFFF5F5F5)),
-            contentScale = ContentScale.Companion.Crop
-        )
-
-        Spacer(modifier = Modifier.Companion.height(12.dp))
-
-        Text(
-            text = product.name,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black,
-            lineHeight = 14.sp
-        )
-
-        Spacer(modifier = Modifier.Companion.height(4.dp))
-
-        Text(
-            text = product.price,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black
-        )
-    }
-}
-
-@Composable
-fun ShopScreen() {
-    var selectedTabIndex by remember { mutableIntStateOf(0) }
-    val tabs = listOf("전체", "Tops & T-shirts", "sale")
-
-    Column(
-        modifier = Modifier.Companion
-            .fillMaxSize()
-            .background(Color.Companion.White)
-    ) {
-        Spacer(modifier = Modifier.Companion.height(28.dp))
-
-        TabRow(
-            selectedTabIndex = selectedTabIndex,
-            containerColor = Color.Companion.White,
-            contentColor = Color.Companion.Black
-        ) {
-            tabs.forEachIndexed { index, title ->
-                Tab(
-                    selected = selectedTabIndex == index,
-                    onClick = { selectedTabIndex = index },
-                    text = {
-                        Text(
-                            text = title,
-                            fontSize = 12.sp
-                        )
-                    }
-                )
-            }
-        }
-
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(2),
-            modifier = Modifier.Companion.fillMaxSize(),
-            contentPadding = PaddingValues(horizontal = 24.dp, vertical = 24.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp),
-            verticalArrangement = Arrangement.spacedBy(36.dp)
-        ) {
-            items(
-                items = productList,
-                key = { product -> product.id }
-            ) { product ->
-                ShopProductCard(product = product)
-            }
-        }
-    }
-}
-
-@Composable
-fun ShopProductCard(
-    product: ProductItem
-) {
-    Column(
-        modifier = Modifier.Companion.fillMaxWidth()
-    ) {
-        Box(
-            modifier = Modifier.Companion.fillMaxWidth()
-        ) {
-            Image(
-                painter = painterResource(id = product.imageResId),
-                contentDescription = product.name,
-                modifier = Modifier.Companion
-                    .fillMaxWidth()
-                    .aspectRatio(1f)
-                    .background(Color(0xFFF5F5F5)),
-                contentScale = ContentScale.Companion.Crop
-            )
-
-            Icon(
-                painter = painterResource(
-                    id = if (product.isLiked) R.drawable.ic_heart_on else R.drawable.ic_heart_off
-                ),
-                contentDescription = "좋아요",
-                modifier = Modifier.Companion
-                    .align(Alignment.Companion.TopEnd)
-                    .padding(8.dp)
-                    .size(22.dp),
-                tint = Color.Companion.Unspecified
-            )
-        }
-
-        Spacer(modifier = Modifier.Companion.height(8.dp))
-
-        product.badge?.let {
-            Text(
-                text = it,
-                fontSize = 11.sp,
-                color = Color(0xFFFF6A00),
-                fontWeight = FontWeight.Companion.Bold
-            )
-        }
-
-        Text(
-            text = product.name,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black,
-            lineHeight = 14.sp
-        )
-
-        Text(
-            text = product.subTitle,
-            fontSize = 10.sp,
-            color = Color.Companion.Gray,
-            lineHeight = 13.sp
-        )
-
-        Text(
-            text = product.colors,
-            fontSize = 10.sp,
-            color = Color.Companion.Gray,
-            lineHeight = 13.sp
-        )
-
-        Spacer(modifier = Modifier.Companion.height(2.dp))
-
-        Text(
-            text = product.price,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black
-        )
-    }
-}
-
-@Composable
-fun WishlistScreen() {
-    Column(
-        modifier = Modifier.Companion
-            .fillMaxSize()
-            .background(Color.Companion.White)
-    ) {
-        Text(
-            text = "위시리스트",
-            fontSize = 24.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black,
-            modifier = Modifier.Companion.padding(start = 24.dp, top = 40.dp)
-        )
-
-        LazyVerticalGrid(
-            columns = GridCells.Fixed(2),
-            modifier = Modifier.Companion.fillMaxSize(),
-            contentPadding = PaddingValues(horizontal = 24.dp, vertical = 28.dp),
-            horizontalArrangement = Arrangement.spacedBy(16.dp),
-            verticalArrangement = Arrangement.spacedBy(28.dp)
-        ) {
-            items(
-                items = productList,
-                key = { product -> product.id }
-            ) { product ->
-                WishlistProductCard(product = product)
-            }
-        }
-    }
-}
-
-@Composable
-fun WishlistProductCard(
-    product: ProductItem
-) {
-    Column(
-        modifier = Modifier.Companion.fillMaxWidth()
-    ) {
-        Image(
-            painter = painterResource(id = product.imageResId),
-            contentDescription = product.name,
-            modifier = Modifier.Companion
-                .fillMaxWidth()
-                .aspectRatio(1f)
-                .background(Color(0xFFF5F5F5)),
-            contentScale = ContentScale.Companion.Crop
-        )
-
-        Spacer(modifier = Modifier.Companion.height(8.dp))
-
-        Text(
-            text = product.name,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black,
-            lineHeight = 14.sp
-        )
-
-        Text(
-            text = product.price,
-            fontSize = 11.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black
-        )
-
-        if (product.id == 2) {
-            Text(
-                text = product.subTitle,
-                fontSize = 10.sp,
-                color = Color.Companion.Gray,
-                lineHeight = 13.sp
-            )
-
-            Text(
-                text = product.colors,
-                fontSize = 10.sp,
-                color = Color.Companion.Gray,
-                lineHeight = 13.sp
-            )
-
-            Text(
-                text = "US$10",
-                fontSize = 11.sp,
-                fontWeight = FontWeight.Companion.Bold,
-                color = Color.Companion.Black
-            )
-        }
-    }
-}
-
-@Composable
-fun CartScreen(
-    onOrderClick: () -> Unit
-) {
-    Column(
-        modifier = Modifier.Companion
-            .fillMaxSize()
-            .background(Color.Companion.White)
-            .padding(horizontal = 24.dp)
-    ) {
-        Box(
-            modifier = Modifier.Companion
-                .fillMaxWidth()
-                .weight(1f),
-            contentAlignment = Alignment.Companion.Center
-        ) {
-            Column(
-                horizontalAlignment = Alignment.Companion.CenterHorizontally
-            ) {
-                Image(
-                    painter = painterResource(id = R.drawable.ic_cartcircle),
-                    contentDescription = "장바구니 아이콘",
-                    modifier = Modifier.Companion.size(56.dp)
-                )
-
-                Spacer(modifier = Modifier.Companion.height(12.dp))
-
-                Text(
-                    text = "장바구니가 비어 있습니다.\n제품을 추가하면 여기에 표시됩니다.",
-                    fontSize = 12.sp,
-                    color = Color.Companion.Black,
-                    textAlign = TextAlign.Companion.Center,
-                    lineHeight = 18.sp
-                )
-            }
-        }
-
-        Button(
-            onClick = onOrderClick,
-            modifier = Modifier.Companion
-                .fillMaxWidth()
-                .height(56.dp),
-            colors = ButtonDefaults.buttonColors(
-                containerColor = Color.Companion.Black,
-                contentColor = Color.Companion.White
-            )
-        ) {
-            Text(
-                text = "주문하기",
-                fontSize = 15.sp,
-                fontWeight = FontWeight.Companion.Bold
-            )
-        }
-
-        Spacer(modifier = Modifier.Companion.height(16.dp))
-        Spacer(modifier = Modifier.Companion.windowInsetsBottomHeight(WindowInsets.Companion.navigationBars))
-    }
-}
-
-@Composable
-fun ProfileScreen() {
-    Column(
-        modifier = Modifier.Companion
-            .fillMaxSize()
-            .background(Color.Companion.White)
-            .padding(horizontal = 24.dp, vertical = 32.dp)
-    ) {
-        Text(
-            text = "프로필",
-            fontSize = 24.sp,
-            fontWeight = FontWeight.Companion.Bold,
-            color = Color.Companion.Black
-        )
     }
 }
